+++ src/constants/index.ts
@@ -114,7 +114,10 @@ export const ALLOW_SIGNUP = process.env.REACT_APP_ALLOW_SIGNUP === "true";
 export const RECAPTCHA_SITE_KEY = process.env.REACT_APP_RECAPTCHA_SITE_KEY;
 
 export const APP_ENVIROMENT = process.env.REACT_APP_ENV || "PROD";
-export const ALLOW_PURCHASE = process.env.REACT_APP_ALLOW_PURCHASE || true;
+export const ALLOW_PURCHASE =
+  process.env.REACT_APP_ALLOW_PURCHASE === undefined
+    ? true
+    : process.env.REACT_APP_ALLOW_PURCHASE;
 
 export enum GalleryContext {
   BACKGROUND,